src/main/java/com/cafe/coco/repository/MemoryMemberRepository.java
package com.cafe.coco.repository;

import com.cafe.coco.domain.Member;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Optional;

public class MemoryMemberRepository implements MemberRepository{
    private static HashMap<Long, Member> store = new HashMap<Long, Member>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setNum(++sequence);
        store.put(member.getNum(), member);
        return member;
    }

    @Override
    public Optional<Member> findByNum(Long num) {
        return Optional.ofNullable(store.get(num));
    }

    @Override
    public Optional<Member> findById(String id) {
        return store.values().stream()
                .filter(member -> member.getId().equals(id))
                .findAny();
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }

}
